Upgrade to autosar-data 0.15, pyo3-0.22
Improvements from this upgrade:
- New API: elements_dfs_with_max_depth()
- Optimize the element insertion code, which dramatically speeds up adding may sub-elements to a single parent
- Loading more than one file into the model could hang in the model merging code; this has been fixed
- The XML specification allows attributes to be enclosed in single quotes. Support for this case has been added
- some error messages contain more details

+++ src/specification.rs
@@ -203,6 +203,6 @@ fn character_data_spec_to_object(spec: &CharacterDataSpec) -> PyObject {
         CharacterDataSpec::UnsignedInteger => Py::new(py, CharacterDataTypeUnsignedInt(()))
             .unwrap()
             .into_py(py),
-        CharacterDataSpec::Double => Py::new(py, CharacterDataTypeFloat(())).unwrap().into_py(py),
+        CharacterDataSpec::Float => Py::new(py, CharacterDataTypeFloat(())).unwrap().into_py(py),
     })
 }
